GitHub seed (201308).

+++ Single-Player Source/dmc/pm_shared/pm_defs.h
@@ -24,6 +24,7 @@
 #define PM_TRACELINE_ANYVISIBLE		0
 #define PM_TRACELINE_PHYSENTSONLY	1
 
+#include "archtypes.h"     // DAL
 #include "pm_info.h"
 
 // PM_PlayerTrace results.
@@ -190,7 +191,7 @@ struct playermove_s
 	int				(*PM_HullPointContents) ( struct hull_s *hull, int num, vec3_t p);   
 	pmtrace_t		(*PM_PlayerTrace) (vec3_t start, vec3_t end, int traceFlags, int ignore_pe );
 	struct pmtrace_s *(*PM_TraceLine)( float *start, float *end, int flags, int usehulll, int ignore_pe );
-	long			(*RandomLong)( long lLow, long lHigh );
+	int32			(*RandomLong)( int32 lLow, int32 lHigh );
 	float			(*RandomFloat)( float flLow, float flHigh );
 	int				(*PM_GetModelType)( struct model_s *mod );
 	void			(*PM_GetModelBounds)( struct model_s *mod, vec3_t mins, vec3_t maxs );
